fix(kafka_client): catch KafkaJSNumberOfRetriesExceeded from new KafkaJS()
new KafkaJS() has its own retry routine integrated and throws KafkaJSNumberOfRetriesExceeded when
only 5 retries of 300ms exceeded. This Exception was not catched and also canceled our wrapped
retry.attmept. Increasing the initialRetryTime and the number of retries could already solve the
issue. However, we wrap a forever retry for just in case.