fix(required_fields): it should not be required to have requiredFields mapped to resourceName
like bufferedFields and timestampFields come in as a single object per resource so should
requiredFields. Why should one provied a complete map of requiredFields but then only the entry for
one resource is used. I keep it backward compatible but in the future requiredFields is expected to
be a decicated object